Extract common util for quota adapters. (#484)
* Extract common util for quota adapters.

* Separate util package.
