+++ adapter/memQuota/BUILD
@@ -10,6 +10,7 @@ go_library(
     ],
     deps = [
         "//adapter/memQuota/config:go_default_library",
+        "//adapter/memQuota/util:go_default_library",
         "//pkg/adapter:go_default_library",
         "//pkg/pool:go_default_library",
         "@com_github_gogo_protobuf//types:go_default_library",
